Re-work VPC KCL
This commit largely reworks the KCL script to remove some of the
complexity, particularly around handling the peering connection route
table entries which were over-complex, hard to read and largely invalid.

It achieves this by simplifying the looping and adding helper functions
to check for whether public route tables are to be included. An
additional "check_route_tables.bash" script is included which reads the
peering connections and traces the routes back to the relevant route
tables.

Subnets inside a subnet set are now additionally tagged with
`giantswarm.io/subnetset: <index>` so they can be targetted wit the
network discovery function and the KCL script is more clearly commented
to try and explain what each section is trying to achieve.