Add RDS Base
RDS Base is a composition that's intended to build any combination of
RDS cluster or instances that act together as a cluster.

This is a KCL based composition for providing databases. Whilst the
overall number of options may seem daunting, most can be set at the top
level for sharing across instances, or simply ignored, falling back to
sensible defaults where relevant